services/images-all/requirements.txt
Flask>=3.0.0
Flask-CORS==4.0.0
Pillow>=10.0.0
gunicorn>=21.2.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
cairosvg>=2.7.0
pillow-heif>=0.13.0
rawpy>=0.18.0